# specs/tsp-output/@typespec/openapi3/generated/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: Comprehensive API for managing beauty salon operations including salons, staff, services, customers, reservations, bookings, and reviews.
  version: 0.0.0
...
paths:
...
  /bookings/{id}/reservations:
    post:
      operationId: BookingOperations_addReservation
      summary: Add reservation to booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.BookingId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Booking'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reservationId:
                  $ref: '#/components/schemas/Models.ReservationId'
              required:
                - reservationId
  /bookings/{id}/reservations/{reservationId}:
    delete:
      operationId: BookingOperations_removeReservation
      summary: Remove reservation from booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.BookingId'
        - name: reservationId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReservationId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Booking'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Bookings
...
  /customers/{id}/reservations:
    get:
      operationId: CustomerOperations_getReservations
      summary: Get customer reservations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.CustomerId'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.ReservationStatus'
          explode: false
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - $ref: '#/components/parameters/Models.PaginationParams.limit'
        - $ref: '#/components/parameters/Models.PaginationParams.offset'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total
                  - limit
                  - offset
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Models.ReservationDetail'
                  total:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  offset:
                    type: integer
                    format: int32
      tags:
        - Customers
  /reservations:
    get:
      operationId: ReservationOperations_list
      summary: List reservations
      parameters:
        - name: salonId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.SalonId'
          explode: false
        - name: customerId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.CustomerId'
          explode: false
        - name: staffId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.StaffId'
          explode: false
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.ReservationStatus'
          explode: false
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - $ref: '#/components/parameters/Models.PaginationParams.limit'
        - $ref: '#/components/parameters/Models.PaginationParams.offset'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total
                  - limit
                  - offset
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Models.ReservationDetail'
                  total:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  offset:
                    type: integer
                    format: int32
      tags:
        - Reservations
    post:
      operationId: ReservationOperations_create
      summary: Create reservation
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          headers:
            location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Reservation'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.CreateReservationRequest'
  /reservations/{id}:
    get:
      operationId: ReservationOperations_get
      summary: Get reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReservationId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.ReservationDetail'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reservations
    patch:
      operationId: ReservationOperations_update
      summary: Update reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReservationId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Reservation'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.UpdateReservationRequest'
  /reservations/{id}/cancel:
    post:
      operationId: ReservationOperations_cancel
      summary: Cancel reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReservationId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Reservation'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
  /reservations/{id}/complete:
    post:
      operationId: ReservationOperations_complete
      summary: Complete reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReservationId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Reservation'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reservations
...
components:
  parameters:
    Models.PaginationParams.limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 20
      explode: false
    Models.PaginationParams.offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 0
      explode: false
  schemas:
...
    Models.Booking:
      type: object
      required:
        - id
        - salonId
        - customerId
        - reservationIds
        - status
        - totalAmount
        - finalAmount
        - paymentStatus
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Models.BookingId'
        salonId:
          $ref: '#/components/schemas/Models.SalonId'
        customerId:
          $ref: '#/components/schemas/Models.CustomerId'
        reservationIds:
          type: array
          items:
            $ref: '#/components/schemas/Models.ReservationId'
        status:
          $ref: '#/components/schemas/Models.BookingStatus'
        totalAmount:
          type: integer
          format: int32
        discountAmount:
          type: integer
          format: int32
        finalAmount:
          type: integer
          format: int32
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
...
    Models.BookingId:
      type: string
...
    Models.CustomerId:
      type: string
...
    Models.Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
    Models.Reservation:
      type: object
      required:
        - id
        - salonId
        - customerId
        - staffId
        - serviceId
        - startTime
        - endTime
        - status
        - totalAmount
        - isPaid
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Models.ReservationId'
        salonId:
          $ref: '#/components/schemas/Models.SalonId'
        customerId:
          $ref: '#/components/schemas/Models.CustomerId'
        staffId:
          $ref: '#/components/schemas/Models.StaffId'
        serviceId:
          $ref: '#/components/schemas/Models.ServiceId'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/Models.ReservationStatus'
        notes:
          type: string
        totalAmount:
          type: integer
          format: int32
        depositAmount:
          type: integer
          format: int32
        isPaid:
          type: boolean
        cancellationReason:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    Models.ReservationDetail:
      type: object
      required:
        - customerName
        - staffName
        - serviceName
        - serviceCategory
        - serviceDuration
      properties:
        customerName:
          type: string
        staffName:
          type: string
        serviceName:
          type: string
        serviceCategory:
          $ref: '#/components/schemas/Models.ServiceCategory'
        serviceDuration:
          type: integer
          format: int32
      allOf:
        - $ref: '#/components/schemas/Models.Reservation'
    Models.ReservationId:
      type: string
    Models.ReservationStatus:
      type: string
      enum:
        - pending
        - confirmed
        - cancelled
        - completed
        - no_show
...
    Models.SalonId:
      type: string
...
    Models.ServiceId:
      type: string
...
    Models.StaffId:
      type: string
...
    Models.UpdateReservationRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Models.ReservationStatus'
        notes:
          type: string
        startTime:
          type: string
          format: date-time
        staffId:
          $ref: '#/components/schemas/Models.StaffId'
      description: Reservation update request with optional fields for partial updates
...
servers:
  - url: https://api.beauty-salon.com
    description: Beauty Salon API
    variables: {}